code review: move tests and function, add jax tests, update docstring
