+++ tests/math/test_binary_linalg.py
@@ -17,6 +17,8 @@ import pytest
 
 from pennylane import math
 
+jnp = pytest.importorskip("jax.numpy")
+
 
 def _is_binary(x: np.ndarray) -> bool:
     """Return whether all entries of a numpy array are binary."""
@@ -121,6 +123,84 @@ class TestIntToBinary:
         assert _is_binary(out)
         assert out.shape == (*x.shape, n)
         assert np.allclose(out, expected)
+
+
+class TestBinaryDecimals:
+    """Test ``qml.math.binary_decimals``."""
+
+    @pytest.mark.parametrize(
+        "phi, p, expected, mod",
+        [
+            (1 / 2, 2, [1, 0], 1),
+            (1 / 2 * 4 * np.pi, 2, [1, 0], 4 * np.pi),
+            (1 / 2, 3, [1, 0, 0], 1.0),
+            (1 / 2 * 7.0, 3, [1, 0, 0], 7.0),
+            ((1 / 2 + 1 / 8 + 1 / 16 + 1 / 32), 2, [1, 1], 1.0),
+            ((1 / 2 + 1 / 8 + 1 / 16 + 1 / 32), 3, [1, 1, 0], 1.0),
+            ((1 / 2 + 1 / 8 + 1 / 16 + 1 / 32), 5, [1, 0, 1, 1, 1], 1.0),
+        ],
+    )
+    def test_binary_decimals_scalar(self, phi, p, expected, mod):
+        """Test that the binary representation or approximation of the angle is correct"""
+
+        assert np.array_equal(expected, math.binary_decimals(phi, p, mod))
+
+    @pytest.mark.parametrize(
+        "phi, p, expected",
+        [
+            (np.array([1 / 2, 1 / 2 + 1 / 4, 1 / 4]), 2, [[1, 0], [1, 1], [0, 1]]),
+            (np.array([1 / 2, 1 / 8, 1 / 4 + 1 / 8]), 3, [[1, 0, 0], [0, 0, 1], [0, 1, 1]]),
+            (
+                np.array(
+                    [
+                        1 / 2 + 1 / 4 + 0 / 8 + 1 / 16 + 1 / 32,
+                        1 / 2 + 1 / 4 + 1 / 8 + 1 / 16 + 1 / 32,
+                        1 / 2 + 0 / 4 + 1 / 8 + 0 / 16 + 1 / 32,
+                        0 / 2 + 1 / 4 + 1 / 8 + 0 / 16 + 1 / 32,
+                        0 / 2 + 0 / 4 + 0 / 8 + 0 / 16 + 1 / 32,
+                        1 / 2 + 0 / 4 + 1 / 8 + 1 / 16 + 1 / 32,
+                    ]
+                ),
+                2,
+                [[1, 1], [0, 0], [1, 1], [1, 0], [0, 0], [1, 1]],
+            ),
+        ],
+    )
+    def test_binary_decimals_array(self, phi, p, expected):
+        """Test that the binary representation or approximation of the angle is correct"""
+        out = math.binary_decimals(phi, p)
+        assert np.array_equal(expected, out), f"\n{expected}\n{out}"
+
+    @pytest.mark.jax
+    @pytest.mark.parametrize(
+        "phi, p, expected",
+        [
+            (jnp.array([1 / 2, 1 / 2 + 1 / 4, 1 / 4]), 2, jnp.array([[1, 0], [1, 1], [0, 1]])),
+            (
+                jnp.array([1 / 2, 1 / 8, 1 / 4 + 1 / 8]),
+                3,
+                jnp.array([[1, 0, 0], [0, 0, 1], [0, 1, 1]]),
+            ),
+            (
+                jnp.array(
+                    [
+                        1 / 2 + 1 / 4 + 0 / 8 + 1 / 16 + 1 / 32,
+                        1 / 2 + 1 / 4 + 1 / 8 + 1 / 16 + 1 / 32,
+                        1 / 2 + 0 / 4 + 1 / 8 + 0 / 16 + 1 / 32,
+                        0 / 2 + 1 / 4 + 1 / 8 + 0 / 16 + 1 / 32,
+                        0 / 2 + 0 / 4 + 0 / 8 + 0 / 16 + 1 / 32,
+                        1 / 2 + 0 / 4 + 1 / 8 + 1 / 16 + 1 / 32,
+                    ]
+                ),
+                2,
+                jnp.array([[1, 1], [0, 0], [1, 1], [1, 0], [0, 0], [1, 1]]),
+            ),
+        ],
+    )
+    def test_binary_decimals_jax(self, phi, p, expected):
+        """Test that the binary representation or approximation of the angle is correct for jax arrays"""
+        out = math.binary_decimals(phi, p)
+        assert jnp.array_equal(expected, out), f"\n{expected}\n{out}"
 
 
 class TestBinaryFiniteReducedRowEchelon:
